Perception types and ops to easynav_sensors
