+++ easynav_common/package.xml
@@ -20,16 +20,9 @@
 
   <depend>rclcpp</depend>
   <depend>rclcpp_lifecycle</depend>
-  <depend>pcl_conversions</depend>
-  <depend>pcl_ros</depend>
   <depend>tf2_ros</depend>
   <depend>geometry_msgs</depend>
-  <depend>sensor_msgs</depend>
-  <depend>vision_msgs</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>nav_msgs</depend>
   <depend>yaets</depend>
-  <depend>cv_bridge</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
