+++ easynav_sensors/CMakeLists.txt
@@ -12,7 +12,14 @@ find_package(rclcpp REQUIRED)
 find_package(rclcpp_lifecycle REQUIRED)
 find_package(easynav_common REQUIRED)
 find_package(tf2_ros REQUIRED)
+find_package(tf2_geometry_msgs REQUIRED)
 find_package(sensor_msgs REQUIRED)
+find_package(vision_msgs REQUIRED)
+find_package(geometry_msgs REQUIRED)
+find_package(nav_msgs REQUIRED)
+find_package(PCL REQUIRED COMPONENTS common io)
+find_package(pcl_conversions REQUIRED)
+find_package(cv_bridge REQUIRED)
 
 set(dependencies
   rclcpp::rclcpp
@@ -28,17 +35,30 @@ set(package_dependencies
 
 add_library(${PROJECT_NAME} SHARED
   src/easynav_sensors/SensorsNode.cpp
+  src/easynav_sensors/types/ImagePerception.cpp
+  src/easynav_sensors/types/PointPerception.cpp
+  src/easynav_sensors/types/IMUPerception.cpp
+  src/easynav_sensors/types/GNSSPerception.cpp
+  src/easynav_sensors/types/DetectionsPerception.cpp
 )
 target_include_directories(${PROJECT_NAME} PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include/${PROJECT_NAME}>
+  ${PCL_INCLUDE_DIRS}
 )
 target_link_libraries(${PROJECT_NAME} PUBLIC
   rclcpp::rclcpp
   rclcpp_lifecycle::rclcpp_lifecycle
   easynav_common::easynav_common
   tf2_ros::tf2_ros
+  tf2_geometry_msgs::tf2_geometry_msgs
+  pcl_conversions::pcl_conversions
+  cv_bridge::cv_bridge
   ${sensor_msgs_TARGETS}
+  ${vision_msgs_TARGETS}
+  ${geometry_msgs_TARGETS}
+  ${nav_msgs_TARGETS}
+  ${PCL_LIBRARIES}
 )
 
 install(
@@ -75,6 +95,13 @@ ament_export_dependencies(
   rclcpp_lifecycle
   easynav_common
   tf2_ros
+  tf2_geometry_msgs
   sensor_msgs
+  vision_msgs
+  geometry_msgs
+  nav_msgs
+  pcl_conversions
+  cv_bridge
+  PCL
 )
 ament_package()
