+++ easynav_sensors/include/easynav_sensors/types/DetectionsPerception.hpp
@@ -19,8 +19,8 @@
 /// and the DetectionsPerceptionsHandler class, which handles subscriptions to image messages and transforms them into
 /// DetectionsPerceptions instances. It also defines an alias for a collection of such perceptions.
 
-#ifndef EASYNAV_COMMON_TYPES__DETECTIONSPERCEPTIONS_HPP_
-#define EASYNAV_COMMON_TYPES__DETECTIONSPERCEPTIONS_HPP_
+#ifndef EASYNAV_SENSORS_TYPES__DETECTIONSPERCEPTIONS_HPP_
+#define EASYNAV_SENSORS_TYPES__DETECTIONSPERCEPTIONS_HPP_
 
 #include <string>
 #include <vector>
@@ -29,7 +29,7 @@
 
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
 
-#include "easynav_common/types/Perceptions.hpp"
+#include "easynav_sensors/types/Perceptions.hpp"
 
 namespace easynav
 {
@@ -111,4 +111,4 @@ rclcpp::Time get_latest_detections_perceptions_stamp(const DetectionsPerceptions
 
 }  // namespace easynav
 
-#endif  // EASYNAV_COMMON_TYPES__DETECTIONSPERCEPTIONS_HPP_
+#endif  // EASYNAV_SENSORS_TYPES__DETECTIONSPERCEPTIONS_HPP_
