+++ easynav_sensors/src/easynav_sensors/types/DetectionsPerception.cpp
@@ -16,57 +16,48 @@
 #include <string>
 
 #include "cv_bridge/cv_bridge.hpp"
-#include "sensor_msgs/msg/image.hpp"
+#include "vision_msgs/msg/detection3_d_array.hpp"
 
 #include "rclcpp/time.hpp"
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
 
-#include "easynav_common/types/ImagePerception.hpp"
+#include "easynav_sensors/types/DetectionsPerception.hpp"
 
 namespace easynav
 {
 
 
 rclcpp::SubscriptionBase::SharedPtr
-ImagePerceptionHandler::create_subscription(
+DetectionsPerceptionsHandler::create_subscription(
   rclcpp_lifecycle::LifecycleNode & node,
   const std::string & topic,
   const std::string & type,
   std::shared_ptr<PerceptionBase> target,
   rclcpp::CallbackGroup::SharedPtr cb_group)
 {
-  if (type != "sensor_msgs/msg/Image") {
-    throw std::runtime_error("Unsupported message type for ImagePerceptionHandler: " + type);
+  if (type != "vision_msgs/msg/Detection3DArray") {
+    throw std::runtime_error("Unsupported message type for DetectionsPerceptionsHandler: " + type);
   }
 
   auto options = rclcpp::SubscriptionOptions();
   options.callback_group = cb_group;
 
-  return node.create_subscription<sensor_msgs::msg::Image>(
+  return node.create_subscription<vision_msgs::msg::Detection3DArray>(
     topic, rclcpp::QoS(1),
-    [target](const sensor_msgs::msg::Image::SharedPtr msg)
+    [target](const vision_msgs::msg::Detection3DArray::SharedPtr msg)
     {
-      auto typed_target = std::dynamic_pointer_cast<ImagePerception>(target);
+      auto typed_target = std::dynamic_pointer_cast<DetectionsPerception>(target);
 
       typed_target->stamp = msg->header.stamp;
       typed_target->frame_id = msg->header.frame_id;
       typed_target->new_data = true;
 
-      try {
-        cv_bridge::CvImageConstPtr cv_ptr = cv_bridge::toCvShare(msg, msg->encoding);
-        typed_target->data = cv_ptr->image.clone();  // se clona para evitar compartir buffers
-        typed_target->valid = true;
-      } catch (const cv_bridge::Exception & e) {
-        RCLCPP_WARN(
-          rclcpp::get_logger("ImagePerceptionHandler"),
-          "cv_bridge exception: %s", e.what());
-        typed_target->valid = false;
-      }
-    },
-    options);
+      typed_target->data = *msg;  // Copy the Detection3DArray message
+      typed_target->valid = true;
+    }, options);
 }
 
-rclcpp::Time get_latest_image_perceptions_stamp(const ImagePerceptions & perceptions)
+rclcpp::Time get_latest_detections_perceptions_stamp(const DetectionsPerceptions & perceptions)
 {
   rclcpp::Time latest_stamp;
   bool inited = false;
@@ -77,7 +68,6 @@ rclcpp::Time get_latest_image_perceptions_stamp(const ImagePerceptions & percept
       inited = true;
     }
   }
-
   return latest_stamp;
 }
