+++ easynav_sensors/src/easynav_sensors/types/ImagePerception.cpp
@@ -15,48 +15,58 @@
 
 #include <string>
 
-#include "sensor_msgs/msg/imu.hpp"
+#include "cv_bridge/cv_bridge.hpp"
+#include "sensor_msgs/msg/image.hpp"
 
 #include "rclcpp/time.hpp"
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
 
-#include "easynav_common/types/IMUPerception.hpp"
+#include "easynav_sensors/types/ImagePerception.hpp"
 
 namespace easynav
 {
 
 
 rclcpp::SubscriptionBase::SharedPtr
-IMUPerceptionHandler::create_subscription(
+ImagePerceptionHandler::create_subscription(
   rclcpp_lifecycle::LifecycleNode & node,
   const std::string & topic,
   const std::string & type,
   std::shared_ptr<PerceptionBase> target,
   rclcpp::CallbackGroup::SharedPtr cb_group)
 {
-  if (type != "sensor_msgs/msg/Imu") {
-    throw std::runtime_error("Unsupported message type for IMUPerceptionHandler: " + type);
+  if (type != "sensor_msgs/msg/Image") {
+    throw std::runtime_error("Unsupported message type for ImagePerceptionHandler: " + type);
   }
 
   auto options = rclcpp::SubscriptionOptions();
   options.callback_group = cb_group;
 
-  return node.create_subscription<sensor_msgs::msg::Imu>(
+  return node.create_subscription<sensor_msgs::msg::Image>(
     topic, rclcpp::QoS(1),
-    [target](const sensor_msgs::msg::Imu::SharedPtr msg)
+    [target](const sensor_msgs::msg::Image::SharedPtr msg)
     {
-      auto typed_target = std::dynamic_pointer_cast<IMUPerception>(target);
+      auto typed_target = std::dynamic_pointer_cast<ImagePerception>(target);
 
       typed_target->stamp = msg->header.stamp;
       typed_target->frame_id = msg->header.frame_id;
       typed_target->new_data = true;
-      typed_target->data = *msg;
-      typed_target->valid = true;
+
+      try {
+        cv_bridge::CvImageConstPtr cv_ptr = cv_bridge::toCvShare(msg, msg->encoding);
+        typed_target->data = cv_ptr->image.clone();  // se clona para evitar compartir buffers
+        typed_target->valid = true;
+      } catch (const cv_bridge::Exception & e) {
+        RCLCPP_WARN(
+          rclcpp::get_logger("ImagePerceptionHandler"),
+          "cv_bridge exception: %s", e.what());
+        typed_target->valid = false;
+      }
     },
     options);
 }
 
-rclcpp::Time get_latest_imu_perceptions_stamp(const IMUPerceptions & perceptions)
+rclcpp::Time get_latest_image_perceptions_stamp(const ImagePerceptions & perceptions)
 {
   rclcpp::Time latest_stamp;
   bool inited = false;
@@ -67,8 +77,8 @@ rclcpp::Time get_latest_imu_perceptions_stamp(const IMUPerceptions & perceptions
       inited = true;
     }
   }
+
   return latest_stamp;
 }
-
 
 }  // namespace easynav
