+++ easynav_sensors/tests/CMakeLists.txt
@@ -1,3 +1,13 @@
+ament_add_gtest(perceptions_tests perceptions_tests.cpp)
+target_link_libraries(perceptions_tests
+  ${PROJECT_NAME}
+)
+
+ament_add_gtest(pointperceptionsops_tests pointperceptionsops_tests.cpp)
+target_link_libraries(pointperceptionsops_tests
+  ${PROJECT_NAME}
+)
+
 ament_add_gtest(sensors_node_tests sensors_node_tests.cpp)
 target_link_libraries(sensors_node_tests
   ${PROJECT_NAME}
